Adding the foundations of reading/writing with JSON files.
Note to self: in the future, all objects should have a template ID that defines the basic features of said object.
Like how the current system uses class prototypes to figure out what values are default and therefore shouldn't be saved,
if a template is set, it will also ignore default values in reference to the template.