Messing with a new concept on command error handling.
Adding channels for handling communication.

Added message categories. Categories are sent to the client.
Instead of the server sending a linebreak to the client if
the last message is a different category, processing will be
handled entirely on the client's side. This allows for more
flexibility on the user interface side.