Working on templates. I like the concept a lot.
A template is basically two parts: the constructor and the
"prototype" of the full object. The prototype is an instance of
the base object type with all of its unique attributes set as
properties on the prototype.

When spawning an instance of a template, it creates a second object
of the same base type, then assigns the "prototype" instance as the
instance's __proto__ attribute.