MASSIVE overhaul of concepts in the source.
I've added a class of objects called managers to avoid cyclical requirements.
What'll happen is the Database class will read files, and then pass them to the
individual managers. The races will go to the race manager, classes to class
manager, etc...

This way, when you want to know who is online, you can just ask for the
PlayerManager instead of asking for the entire framework for loading
every single type of data from the local database (of JSON files).

Not totally done (need to make a reference aka helpfile manager).
This manager will include stuff like the greeting and the MOTD.

Added direction flags and an object for manipulating them.