Added a MapObjectFactory to disambiguate object loading functions from either the Template or the MapObject.
The MapObject can't be allowed to know about its children because
it creates a cyclical reference, which I am trying specifically
to avoid here in this change.

Templates now have names instead of number IDs. You can basically
think of it as a string ID, but I changed it to name cause I felt
like it.

Templates can now refer to objects that refer to other objects.
As of right now it only works for the contents of an object.
In future it can be adapted to do other stuff. :)