MODELS: ------- Added a concept of models, which use templates to create a fully fleshed out instance of an object in the game.
For example, the HobGoblin template will inform the basic parameters
of what a HobGoblin is, while the "Keith the HobGoblin" model will
have things like what equipment he's wearing, what items are in his
inventory, and things like that.

Objects that have a template or a model shouldn't save their
constructor -- the model and template will do that for us.

I've also set it up so models and templates can be stored in
arbitrary sub directores in order to help with organizing different
groups. I'll have to set it up so that when saving a model/template,
it uses the name to determine where to save it. As it stands, the
name of a model/template should indicate the subdirectory tree.
i.e. the Weapon.excalibur template should be saved in
/data/template/Weapon/excalibur. Same for models.

I did this for 2 reasons:
1) It makes sense to have basic types with nothing associated with
   them.

2) It makes it easier to deal with since all basic types will be
   loaded as Templates before I used them in Models. This means
   I don't have to do any extra pre-processing to avoid
   things like loading them out of order.